requirements.txt
torch~=1.8.1
datasets==1.6.2
pytorch-lightning==1.3.3
transformers==4.6.1
hydra-core
wandb
numpy
scikit-learn
python-telegram-bot
natasha
transformers